requirements.txt
transformers
accelerate
peft
torch
opencv-python
numpy
pillow
scikit-image
matplotlib


